cache task & background cache loading/reloading
