+++ src/instrumentation.ts
@@ -154,3 +154,11 @@
 //     }
 
 // }
+
+export async function register() {
+  if (process.env.NEXT_RUNTIME === "nodejs") {
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    const { registerServer } = require("./instrumentation_server") as { registerServer: () => void };
+    registerServer();
+  }
+}
