+++ src/instrumentation_server.ts
@@ -1,0 +1,21 @@
+import cron from "node-cron";
+import { cacheTask } from "./lib/cache";
+
+let cronTask: cron.ScheduledTask | null = null;
+
+export async function registerServer() {
+  if (cronTask === null) {
+    console.log("Starting cache cron job...");
+    await cacheTask();
+    cronTask = cron.schedule("0,5,10,15,20,25,30,35,40,45,50,55 * * * *", () => {
+      console.log("Running cache task", new Date().toISOString());
+      void (async () => {
+        try {
+          await cacheTask();
+        } catch (k) {
+          console.error("cacheTask error", k);
+        }
+      })();
+    });
+  }
+}
